rename models db
